ooptimize

+++ src/com_mem.cpp
@@ -412,7 +412,7 @@ bool com_mem_to_file(std::string file)
         return false;
     }
     std::string json = com_mem_to_json();
-    int ret = com_file_write(file_fd, json.data(), json.size());
+    int ret = com_file_write(file_fd, json.data(), (int)json.size());
     com_file_close(file_fd);
     return (ret == (int)json.size());
 }
